Move the bid count from the title to the contact info box on the details page
