+++ src/sass/_details.scss
@@ -289,7 +289,7 @@
       white-space: nowrap;
     }
   }
-  
+
   .bid-count-container {
     margin-right: 7.5px;
   }
@@ -392,10 +392,22 @@
 
   .contact-container:first-of-type {
     border-bottom: 0;
+    position: relative;
+  }
+
+  .contact-container:nth-of-type(2) {
+    &.has-bid-count {
+      padding-top: 2em;
+    }
+  }
+
+  .bid-count-container {
+    bottom: -12px;
+    position: absolute;
   }
 
   .contact-container:nth-of-type(2):not(.no-button) {
-    padding-bottom: 4em;
+    padding-bottom: 2em;
   }
 
   .offset-bid-button-container {
